refactor(tests): separate engine process integration
